requirements.txt
# Core ML and AI Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
peft>=0.4.0
datasets>=2.12.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.25.0
gradio>=3.35.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Cloud and Deployment
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0
docker>=6.1.0
kubernetes>=26.1.0

# Monitoring and Logging
mlflow>=2.5.0
wandb>=0.15.0
prometheus-client>=0.17.0
structlog>=23.1.0

# Configuration and Utils
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
rich>=13.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development Tools
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Additional Utilities
requests>=2.31.0
aiofiles>=23.1.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Performance Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
tensorrt>=8.6.0
